5 percent price increase added

+++ src/main/java/Objects/Room.java
@@ -3,9 +3,9 @@ package Objects;
 public class Room {
 	public int id;
 	public String size;
-	public int cost;
+	public double cost;
 	
-	public Room(int id, String size, int cost) {
+	public Room(int id, String size, double cost) {
 		this.id = id;
 		this.size = size;
 		this.cost = cost;
